apply SessionActivity.kt 생성
