+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Ifkakao">
         <activity
+            android:name=".presentation.SessionActivity"
+            android:exported="false" />
+        <activity
             android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
