+++ app/src/main/java/com/survivalcoding/ifkakao/di/AppModule.kt
@@ -3,6 +3,8 @@ package com.survivalcoding.ifkakao.di
 import com.survivalcoding.ifkakao.data.datasource.SessionDataSource
 import com.survivalcoding.ifkakao.data.datasource.SessionDataSourceImpl
 import com.survivalcoding.ifkakao.data.datasource.SessionService
+import com.survivalcoding.ifkakao.data.repository.SessionRepositoryImpl
+import com.survivalcoding.ifkakao.domain.repository.SessionRepository
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -11,10 +13,15 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-object DataSourceModule {
+object AppModule {
 
     @Provides
     @Singleton
     fun provideConferenceDataSource(sessionService: SessionService): SessionDataSource =
         SessionDataSourceImpl(sessionService)
+
+    @Provides
+    @Singleton
+    fun provideSessionRepository(sessionDataSource: SessionDataSource): SessionRepository =
+        SessionRepositoryImpl(sessionDataSource)
 }
